Adding vanilla API entrypoint
