+++ packages/framer-motion/rollup.config.js
@@ -3,13 +3,12 @@ import { terser } from "rollup-plugin-terser"
 import replace from "@rollup/plugin-replace"
 import dts from "rollup-plugin-dts"
 import pkg from "./package.json"
-import { visualizer } from "rollup-plugin-visualizer"
 
 const config = {
     input: "lib/index.js",
 }
 
-const replaceSettings = (env) => {
+export const replaceSettings = (env) => {
     const replaceConfig = env
         ? {
               "process.env.NODE_ENV": JSON.stringify(env),
@@ -79,7 +78,7 @@ const umdProd = Object.assign({}, umd, {
 })
 
 const cjs = Object.assign({}, config, {
-    input: ["lib/index.js", "lib/animation/animate.js"],
+    input: ["lib/index.js", "lib/dom-entry.js"],
     output: {
         entryFileNames: `[name].js`,
         dir: "dist/cjs",
@@ -90,8 +89,8 @@ const cjs = Object.assign({}, config, {
     external,
 })
 
-const es = Object.assign({}, config, {
-    input: ["lib/index.js", "lib/animation/animate.js"],
+export const es = Object.assign({}, config, {
+    input: ["lib/index.js", "lib/dom-entry.js"],
     output: {
         entryFileNames: "[name].mjs",
         format: "es",
@@ -101,75 +100,6 @@ const es = Object.assign({}, config, {
     },
     plugins: [resolve(), replaceSettings()],
     external,
-})
-
-const sizePlugins = [
-    resolve(),
-    replaceSettings("production"),
-    terser({ output: { comments: false } }),
-]
-
-const motion = Object.assign({}, es, {
-    input: "lib/render/dom/motion.js",
-    output: Object.assign({}, es.output, {
-        file: `dist/size-rollup-motion.js`,
-        preserveModules: false,
-        dir: undefined,
-    }),
-    plugins: [...sizePlugins],
-    external: ["react", "react-dom"],
-})
-
-const m = Object.assign({}, es, {
-    input: "lib/render/dom/motion-minimal.js",
-    output: Object.assign({}, es.output, {
-        file: `dist/size-rollup-m.js`,
-        preserveModules: false,
-        dir: undefined,
-    }),
-    plugins: [...sizePlugins, visualizer()],
-    external: ["react", "react-dom"],
-})
-
-const sizeAnimate = Object.assign({}, es, {
-    input: "lib/animation/animate.js",
-    output: Object.assign({}, es.output, {
-        file: `dist/size-rollup-animate.js`,
-        preserveModules: false,
-        dir: undefined,
-    }),
-    plugins: [...sizePlugins],
-    external: ["react", "react-dom"],
-})
-
-const domAnimation = Object.assign({}, es, {
-    input: {
-        "size-rollup-dom-animation-m": "lib/render/dom/motion-minimal.js",
-        "size-rollup-dom-animation": "lib/render/dom/features-animation.js",
-    },
-    output: {
-        format: "es",
-        exports: "named",
-        preserveModules: false,
-        entryFileNames: "[name].js",
-        chunkFileNames: "size-rollup-dom-animation-assets.js",
-        dir: `dist`,
-    },
-    plugins: [...sizePlugins],
-    external: ["react", "react-dom"],
-})
-
-const domMax = Object.assign({}, es, {
-    input: {
-        "size-rollup-dom-animation-m": "lib/render/dom/motion-minimal.js",
-        "size-rollup-dom-max": "lib/render/dom/features-max.js",
-    },
-    output: {
-        ...domAnimation.output,
-        chunkFileNames: "size-rollup-dom-max-assets.js",
-    },
-    plugins: sizePlugins,
-    external: ["react", "react-dom"],
 })
 
 const types = {
@@ -182,10 +112,10 @@ const types = {
 }
 
 const animateTypes = {
-    input: "types/animation/animate.d.ts",
+    input: "types/dom-entry.d.ts",
     output: {
         format: "es",
-        file: "dist/animate.d.ts",
+        file: "dist/dom-entry.d.ts",
     },
     plugins: [dts()],
 }
@@ -206,12 +136,7 @@ export default [
     umdProd,
     cjs,
     es,
-    motion,
-    m,
-    domAnimation,
-    domMax,
     types,
-    sizeAnimate,
     animateTypes,
     threeTypes,
 ]
