+++ packages/framer-motion/src/dom-entry.ts
@@ -1,0 +1,29 @@
+export { motionValue, MotionValue, PassiveEffect, Subscriber } from "./value"
+export { animate } from "./animation/animate"
+export { scroll } from "./render/dom/scroll"
+export { inView } from "./render/dom/viewport"
+
+/**
+ * Easing
+ */
+export * from "./easing/anticipate"
+export * from "./easing/back"
+export * from "./easing/circ"
+export * from "./easing/ease"
+export * from "./easing/cubic-bezier"
+
+/**
+ * Utils
+ */
+export { transform } from "./utils/transform"
+export { clamp } from "./utils/clamp"
+export * from "./utils/delay"
+export * from "./utils/distance"
+export * from "./utils/errors"
+export * from "./utils/interpolate"
+export { mix } from "./utils/mix"
+export { pipe } from "./utils/pipe"
+export { progress } from "./utils/progress"
+export { wrap } from "./utils/wrap"
+export { sync } from "./frameloop"
+export { frameData } from "./frameloop/data"
